Added `gm` and `gmm` mappings to move text
In normal mode, `gm` takes two motions. The first is the text on which it
operates (e.g. ap for a paragraph). The second is the motion which is
applied to the text, like 2j to move the block 2 lines down.

In visual mode, `gm` takes only one motion and moves the selection
accordingly.

In normal mode only, `gm` takes only one motion and moves the current
line accordingly.

Note that in normal mode, `gm` and `gmm` alter the selection marks, so
after moving some text, `gv` select the moved text, not the previous
selection.