The list option is now set only in neovim
Let's be honest: it look _really_ ugly in vim, doesn't it ?

If you want to know why I gave up on this option, here's why :

I tried to fix it first by linking SpecialKey to Conceal, but then the
keys in the intro screen were highlighted with Conceal.

Then I tried extracting the highlight properties of the SpecialKey
group, create a new highlight group oSpecialKey with them and adding
8:oSpecialKey to the 'highlight' option before linking SpecialKey to
Conceal.

Again, it didn't work, the list option was using the highlight group of
the special keys so oSpecialKeys, as if I had done nothing.

So I gave up.
This is raging when you know that all of this is just because the _same_
highlight group is used for two things which are *really* different.