vim: Changed `formatoptions`
Yes, after removing the o flag in 'formatoptions', I now add the r flag.
It is responsible for continuing the current comment when enter is hit
from insert mode, which is the behavior I want.
I didn't notice it until recently, because all "well written" ftplugins
files for vim should reset the 'formatoptions' to their default value,
which has the r flag, and the python ftplugin isn't doing that, so I
enforce it myself!