git: Fixed vim themes submodules
If I am correct (which is not so likely), I just needed to add an empty
folder to mark the location of the submodule, even though it exists in
the .gitmodules file.
Anyways, I will read more about git submodules. Now I understand why
they are considered as a complex feature...