fish: Added setup script helper function
At one moment, I got tired of doing `. ~/.config/fish/setup.fish` every
time I wanted to test something in my setup script. So I added a setup
function. That does extra things.

Basically, there is the setup script and the setup function.
The first one is tedious to invoke, you have to type in the whole path.
The second is very easy to invoke, so in order to avoid mistakes (even
though it will probably never happen) it prompts you when you call it
after the first time.