fish: Improved `confed` completion
Now, the command line options are completed and the file completion
works on macOS.
Before, the file completion made use of the -printf option of GNU find,
which is not available in BSD find.

Also, I refactored the two separate completions for regular and special
config file into one function so there is now only one file completion.