vim: Removed 'showbreak' option and its machinery
Actually, this commit was planned, and I have already removed the
'showbreak' option from my vim config quite a while ago, and I can
actually say that this showbreak option was just extra work for nothing!

The main issue with this option was that it appeared *in* the number
column. Picture something like this:

+----+-----------------------------------+
|  1 |An example file with a really long |
|   ↳|line that is long enough to wrap   |
|   ↳|arround twice.                     |
|  4 |                                   |
:  5 :                                   :
'    '                                   '

Yes, this looked fancy.

This was achieved by setting cpo+=n so the line numbers would not appear
on the wrapped lines, set &showbreak to '   ↳' and winhighlight+=NonText:LineNr

But, there is always a but in Vim, in windows that did not have line
numbers (mostly the help window), it showed the showbreak nonetheless,
which looked something like this:

+-------------------------------------+
|An example file with a really long   |
|   ↳line that is long enough to wrap |
|   ↳arround twice.                   |
:                                     :
'                                     '

So I added autocommands to set correctly 'showbreak' when a window was
entered and all that stuff, and the piece of vimscript that was
initially one line kept growing.
But it worked.
Until I switched to neovim.

For no reason at all, in neovim, the 'showbreak' option is not buffer
local. That means that when I entered a window with no number, every
other window started to look like this:

+----+-----------------------------------+
|  1 |An example file with a really long |
|NONE|line that is long enough to wrap   |
|NONE|arround twice!                     |
|  4 |                                   |
:  5 :                                   :
'    '                                   '

(NONE is a marker to indicate an empty 'showbreak' for the buffer local
value, because `setlocal showbreak=` resets it to the global value)

I tried to fix the NONEs and... The other windows looked like this:

+----+-----------------------------------+
|  1 |An example file with a really long |
|line that is long enough to wrap        |
|arround twice!                          |
|  4 |                                   |
:  5 :                                   :
'    '                                   '

The NONE thing looked better.

So at this point I basically gave up a went like "Anyway, who needs
'showbreak' in the numbers column ?"

And I deleted it.