fish: Created `dotadd` and `dotlink` functions
I was tired of typing the full path of everything when I was linking
from and into my dotfiles so... Here are helper functions!

The `dotadd` function extract the path of a file relative to ~ and moves
it to ~/dotfiles/$path (this can be changed through the -d/--dot-home
switch). It then links the file back to its original location with
`dotlink`, though this final step can be disabled with the -m/--no-link
switch.

On the other hand, `dotlink` creates a symlink from a file in the
dotfiles to its corresponding location in the home folder. Again, the
location of the dotfiles can be changed with -d/--dot-home.