fish: prompt: Add custom prompt postfix
This postfix defaults to '-+ ' but changes with the level of shell
nesting, e.g. it will become '--+ ' when you nest fish once and '---+ '
when you nest fish twice.

I forgot to add this function when I added the new prompt (even though
it is not new at all, I am just lagging so far behind with commits...)