vim: syntax: Add custom syntax file for fish
I usually use vim-polyglot for the syntax highlighting, but the syntax
file for fish included with it is just no good enough for me...
So I created my own!

I created it with the idea of being as close to fish's builtin syntax
highlighting as possible, so it had to:
	- Allow to use multiple kind of parts in a token (string and
	  variables), even in the command.
	- Handle well new commands after delimiters (pipes, semicolons
	  and ampersands)
	- Allow command substitution

And I thing that overall, I pretty much nailed it!
Of course, there are some missing features, as i/o redirection
highlighting or wildcards highlighting, but it can wait.