fish: Remove entirely ccutils dependency
This used to be my completion framework... It was basically syntactic
sugar around fish's builtin completion helper.

I removed it because it is not included in my dotfiles, which meant that
some completions didn't work on other computers.

I will not work any longer on ccutils, but I may create a static
completion generator (which will accept files in yaml or something like
that). Don't expect this to come one day, though. I may never even try
to do it.