git: Sort submodules in alphabetical order
Well... Not the themes.

First, there are a whole bunch of them and I wasn't motivated to do it,
and second, I plan to use something else than submodules to manage my
themes, as there are a lot of them and the list keeps growing, and I
want them to be lazily loaded.