vim: Remove `plugin/netrw.vim`
I am removing this file because ... I found out that there was something
in it causing a use of &shell at startup.
Obviously, I don't like this.

I discovered this because it was blocking vim when I was resolving a
merge conflict in my dotfiles. This conflict was disabling me from using
my shell - or more specifically to open new shells. Nonetheless, I
managed to resolve it without too much efforts.

But still, I don't want my shell to be used when I start my editor!

So I did the divide-and-conquer strategy. By this I mean I disabled half
of my config, checked if the bug was still there. If it were, I'd enable
back the half and disable the other one. If it were gone, I'd enable
back half of what I disabled.

In the end, it was this 9-line file. Incredible, right? Yet, it was
using netrw_gitignore#Hide(), which was in turn using system() to get
the list of files to ignore from git.

And because I don't use NetRW, I promptly deleted this file.