vim: Fix indentation related stuff.
By this, I mean:
  1) vim-polyglot's auto indent detection
  2) Set only 'tabstop' in ftplugin/vim.vim

If you wonder why, it is because vim's indentation situation in kind of
a mess. There are four options to set indentation, but only two seem
trivial: 'tabstop', which sets the tabstop, and 'expandtab', which
control whether indendation is done with tabs or spaces.

The two other are 'shiftwidth' and 'softtabstop'.

While writing the previous sentence, I discovered a new one:
'vartabstop'. Isn't it crazy?

'shiftwidth' controls the number of spaces each level of indentation is,
mixing tab and spaces when needed to achieve the desired length.

On the other hand, 'softtabstop' controls the length of the indentation
inserted when a tab is inserted. Again, it will mix tab and spaces as
needed to reach the desired length.

So basically, with these two options, you can keep the standard 8 spaces
tab yet indent using 4 "spaces equivalent". By this, I mean that two
levels of indentation (8 spaces) will be turned into a tab.
That's wild!

However, my indentation needs are much simpler. I want to use either
*one* tab per indentation level when 'expandtab' is off, or use
'tabstop' spaces per indentation level when 'expandtab' is on.
Quite simple, right?

That's why I decided to use just the 'tabstop' and the 'expandtab'
options. Therefore 'shiftwidth' has to always stay at 0 (to use
'tabstop') else I begin to get weird results.

As a bonus, everything scales when I change the 'tabstop', including the
> and < commands and indent guides.

So I try to manage it myself as much as possible.