nvim: Add script to use current neovim instance
This is a script that I wanted to make for a long time, and I made it
some time ago!

The script - that I will call nvim.py here - simply reads
$NVIM_LISTEN_ADDRESS and uses pynvim to tell nvim to edit the files,
open new tabs and more.

There is also a companion script - nvim-wait - that tells nvim.py to
block until the opened buffers are dropped. This script is set as
$EDITOR because it is is expected to block by many programs, including
git. It is not a great experience (I need to do :bdelete # to resume the
execution), but it works.