nvim: Add LSP support!
It's been a long time since I wanted to use LSP servers with neovim
(though I was already using them but my config wasn't tracked so I lost
it all).

And at last, here it is!

This is only the first part because I plan to use cmp (previously compe)
for the autocompletion, though I still have things to figure out...

I'm so excited!