nvim: Changed theme-switching mechanism
This is a major rewrite of `after/plugin/colors.vim`.

I moved to a 'background'-based system, so now everything is driven by
doing `:set background=...`.

The best benefit of this is that I don't have to keep a "reversible
themes" table! This was stupid from the very beginning, since those
themes already support the 'background' option.

Also, I renamed `SetBackground()` function to `UpdateTheme()`.
You shouldn't have to interact with it now, anyway.

The only concern I have is that the theme might be set twice, but who
cares?