nvim: Fix the theme-switching mechanism...
Yes, I've just made a change to the theme-switching mechanism, and it is
already broken. I know.

I pushed the change because I hadn't tested it... What a fool...

The thing that I hadn't accounted for was the events that happen when
one does `:set background=light`.

The problem is that if the current colorscheme isn't made for light and dark themes
using the 'background' option, it will set it on its own. And if it is
the case, vim/neovim resets the value of g:colors_name.

The logic here is that if the `background` is set, it is either a
mistake or not for this theme (i.e. to "prepare" the next theme).

I had to look thoroughly into the process to determine what was causing
this, with several autocommands (ColorSchemePre, OptionSet and
ColorScheme), a custom colorscheme and even `nvim --clean`.

I eventually settled on saying that it was a *feature* of neovim, so I
looked in the source code, and hooray! The line responsible of this is
the line 2534 of src/nvim/options.c [1].

So I introduced an autocommand for ColorScheme, which is triggered
before g:colors_name is unset, to save the value of the variable.

I'm glad it finally works.

[1]: As of writing. See https://github.com/neovim/neovim/blob/30c9c8815b531e0130ebeb9358bc6d3947a6128a/src/nvim/option.c#L2534
     for the specific commit.