fish: Rewrite the `del` function
Before, this function used to be a front end to the OS' `rm` commands,
using the -i or -I flag, but it was limited by the capacities of these
commands.

Now, it filters the -i and -I options and does it on its own, and passes
the remaining arguments to the OS' rm command.

This allows me to emulate the -I flag from the GNU rm on other systems,
such as the BSD, and thus macOS.