fish: Add privileges escalation support to `del`
Isn't it annoying when a file can't be deleted in a `rm -r` command, and
you have to manually clean the leftovers?

Isn't it annoying that with a custom `del` function, sudo !! (or alt+s
the fishy way) doesn't work?

So the privilege escalation is now built in the `del` function!

(This is basically doing `sudo !!` but automatically)