nvim: Add a custom tabline in lua!
It has been a long time since I wanted to have a custom lua tabline.

To be honest, I had already done a lot of lua scripting before but I had
never committed it because I had too much changes to commit and I didn't
want to do it all at once.

This is just a reimplementation of the default tabline with less
features though 😅 : it doesn't show when a file is modified or the
number of opened windows in a tab.

However, this is fine: I don't like the window indicator and I plan to
reimplement the modified indicator.

The current state of the tabline is intended to be only temporary: this
is just a base onto which I'll built something better (devicons, close
button/modified indicator, maybe pinned tabs, and in general whatever I
want).

If I find a plugin that suits me well (that is, fully configurable by
me), I'll probably use it but right now, I've found nothing that is
satisfying enough.