nvim: Extract `tabline/init.lua` to `tabline/tabs.lua`
Yes, more refactoring.

I actually built a prototype of pinned tabs and this refactor was
needed, to have a custom `render()` function that displayed both the
tabline and the pinned tabs line.

In the future, I know that I'll add more stuff to the tabline (like cwd
indicator), so if I hadn't refactored it now, I'd have to refactor this
file sooner or later anyways...