nvim: Add the editorconfig.nvim plugin
This plugin adds support to the technology of (almost) the same name, editorconfig.
See https://editorconfig.org/

There is a similar plugin for vim (editorconfig-vim), but this one
claims to be lighter, and I like this.

Also, it is written in lua, and lua is the future of neovim, as every
angry popup-window-lsp-fan-neovimer would say! (I'm not saying that
popup windows and lsp aren't great, I like them too. It was just a joke)