nvim: Navigate up with `-` in a floating window
This is not the best name for this commit, but I've got to keep it under
50 characters...

Basically, I'm using neovim, so I can't not use floating windows. This
is a *must* for any modern neovim config (half joking half serious).

So my file browser opens in a floating window!

I wanted this feature so much, because I use `-` very often, but like
really often, and having it inside of a floating window seemed like it
was fun...

The "floating window" part of the thing was actually easier than I
expected, and it was basically just figuring out how to call nvim_open_win...

What wasn't easy, though, was opening the files! I quickly realised that
the files were opening *in* the floating window, not in the previous
window. What was I thinking, this is the *expected* behaviour!

So I looked up the help for nerdtree, and I realised that I couldn't do
this easily. So I had to dig into the code of nerdtree, which shows once
again that I should switch to something else.

I eventually found `g:NERDTreeFileNode.GetSelected()`, and all my
problems were solved.