fish: Extract clock colours to their own variable
The clock used to use the colours of the host for the brackets and the
cwd for the time, which was intended to have a "dynamically" themed
clock. That is, it would change its colour with the other colours.

However, I have a computer on which I use a different theme, and the
clock was like *really* ugly because the colours of the host and the cwd
didn't match up well, so I had to use individual variables.

I figured that giving a matching theme in my setup.fish would serve a
similar purpose, because it is like the "default theme" that I ship.
Anyway, anything other than this theme is beyond the scope of *my*
config, so I can't really make assumptions about the colours.