fish: Significantly rewrite `fish_hooks`
The first *big* change is that the hooks that used to be:
 - .fish_hooks, .fish_hooks.d
 - .fish_unhook, .fish_unhook.d
are now:
 - .hook.fish, .hook.fish.d
 - .unhook.fish, .unhook.fish.d

The motivation behind this change is... Editor support! Neovim doesn't
see `.fish_hooks` as a fish file (just as every other editor, it doesn't
end with `.fish`).

Also, I've change the order in which the operations are done, to avoid
getting stuck in a "hook loop" if a hooks happens to do `exec fish`.