nvim: lua: Add extended `vim.validate` helpers
I added `isintance(value, types, nilable = false)` that checks if the
type of `value` is one of `types`.

The function `union(types, optional = false)` is a wrapper around
`isintance` to use it with `vim.validate` easily.

There is also `validate_table(table, spec)`, whose name I don't like at all.
`spec` is a table of the form { [key] = <type> }, with type being
either a string, a function or a table of the form { string | function, message }.

I know this sound complicated, but it really isn't.

And as expected, there is the `table(spec)` function that can be easily
used with `vim.validate`.

I just *love* these helpers, and who knows, maybe I'll try to get them
included in the upstream neovim one day.