fish: Make VCS status in right prompt asynchronous
New feature! And hand-made! (I could've just used a plugin for that,
but... where's the fun in this?)

So the VCS status in now drawn *after* the prompt is drawn, which sounds
a bit weird, because it is. The gist of it is that there is a first
render of the prompt that notices that no CVS status has been rendered,
so it schedules a render of this status. When the render is done -- this
is detected with the --on-job-exit hook of functions -- a second render
of the prompt is scheduled. In this render, the prompt has all of the
resources it needs and the prompt is fully drawn.

I also added a bit of caching onto that so there is no flicker at each
new prompt. The flicker still occurs, though, when you leave a
version-controlled directory.

I really should've used a plugin for that.