fish: Limit to one VCS status job at a time
The issue here was that... If one was to spam the <enter> key, a new
fish job computing the VCS status was spawned every at every key press.
This obviously meant that some limitation was needed.

To achieve this, a flag (with the pid of the current job) is set when a
job is started ($__right_prompt_job_pid) and it is later checked to see
if a job is already running before launching a new one.

The only problem is that if the VCS status takes a *really* long time to
compute (which is why I made it async, remember?), the state of the repo
can change between the moment at which the status is requested and the
moment at which it is delivered.
(if the user can execute a command quickly enough that should modifies
the status shown, this happens)

But really, between an *inaccurate status* for one prompt or being
spammed with fish jobs, the choice is quite obvious to me.