fish: Refactor async computing in its own function
What I mean is that I had made a solution just for the right prompt but
I felt that it was like ounces away from being a general solution.

Most of the things this function does is launching the specified job,
setting a handler for when the job exits and after that putting the
output of the job in a variable.

The new function __kmz_async_compute[1] takes a variable name and a
command (you can quote it or leave it unquoted) and spawns the job
specified by `command`. On completion, it sets the variable `variable`
to the output of the job, so a callback can be specified using
--on-variable.

The point of having a such function instead of asking everyone and their
dog to do it on their own is that it introduces a level of abstraction
and does it in the right way: the function cleans everything up after
the execution

[1]: "kmz" stands for "chromosore" here, I like it better than "cms"