nvim: Fix upwards navigation for non-existing dirs
This affected the files whose name isn't the name of a real file on the
disk and thus the unnamed files.

The problem was that I set up and various settings for the nerdtree file
explorer in the floating window but the guard for non-existing
directories returned early so this configuration isn't done.

Now it only set the edited directory to the pwd and the process can
carry on.