kitty: Fix shell integration
After the last update of kitty, my shell integration was broken because
kitty was behaving *as it should*.

The problem was that before this, it wasn't and setting shell
integration to "disabled" didn't actually do anything. Now that it does,
kitty doesn't set the environment variables that tell that the shell
integration should be enabled anymore, so I had to set it to "no-rc"
instead to have the variables.