nvim: Get rid of `vim-sensible`
This is something that I wanted to do for a long time but I hadn't done
it because I thought that I'd need to move all of vim-sensible's
functionality in my personal configuration.

However, I've learnt recently that a lot of changes had been done to
neovim's defaults ('hidden'...), and some of vim-sensible's part had
been integrated into neovim. So I quickly checked how much of
vim-sensible was already covered by neovim, and turns out... All of it!

So, without further ado, I've gotten rid of vim-sensible. (Also, it was
like a quarter or a half of my startup time so that's great. Checkout
the beautiful `vim-startuptime plugin`__ by dstein64 by the way)

__ http://github.com/dstein64/vim-startuptime