fish & nvim: Add private mode
Basically, I have a folder that contains sensitive stuff (passwords)
(don't worry it is encrypted) and I want to have a kind of "anonymous"
session when I enter this folder, so I set up vim to save absolutely
*nothing*, so no shada file, no undo history, no swap file and no backup.

And, because I'm kind of an `architecture astronaut`__, I thought that I
could encapsulate this behaviour in a function! And there you have a
'privatise' function that you can reuse wherever you want!

Side note: the nice thing about this function is that it automatically
upgrades the session to a private fish session (`fish -P`) and if the
current session is already private, it just executes itself!
This enables the use of this function in custom context, such as a
function that takes care of upgrading the session itself.

__ https://www.joelonsoftware.com/2001/04/21/dont-let-architecture-astronauts-scare-you/