nvim: Add the `:Justify` command
So I had a bit of text that was justified in my browser and I copied it
into neovim and I thought like "how could would it be to have it justified?"

So I made a method to justify the text. Mainly because the
implementation was funny. Basically, I'm expanding the spaces.

I'll describe it here.

I basically find the longest line (unless a length is specified), and
then for each line I find how much character I need to add to fill the
line.

Then, I count the number of spaces in that line and I establish how much
"new spaces" a space will be worth (spacenr + gap / spacenr).

The problem is that this ratio of "new spaces" per space isn't whole. So
I have to distribute the spaces evenly throughout the string. To do
this, I go through every space of the string and I find out by how much
it should be expanded.

This is the trickiest part to explain so I'll say that I keep the deltas
of the floored values of space added so far.
Let me explain: let's say that a space is worth 1.2 new spaces. If I
keep an accumulator, it goes 0, 1.2, 2.4, 3.6, 4.8, 6. And here is the
trick: the delta of ⌊1.2⌋ and ⌊2.4⌋ is 1 but the delta of ⌊4.8⌋ and ⌊6⌋
is 2. So every 5 space there is an extra 6th space added. Repeat that
for all of your string and you distribute the extra space (almost) evenly.

Oh and by the way, the justify function can be used as a 'formatexpr':
:set formatexpr=v:lua.chromosore.commands.justify()