fish: Make `del` prompt to retry as root when unsure
The problem that I was facing was that sometimes when trying to remove a
directory was  that the  directory itself  was writable  but some  files
inside of it  weren't. In that case,  `del` wouldn't  try to delete  the
directory as root because it wouldn't be sure if it was the cause of the
problem.

Now, I can ask a human about  it! The human *knows* what failed (thanks
to the error messages) so it  can make an educated decision about this.
Yay!