fish: Add `pkg-remove` script
This is a tiny yet super useful script that removes macOS pkgs, also
known as "Mac OS X Installer flat packages" (from pkgutil's manpage).

I had this issue earlier of not being able to remove *easily* the files
installed by the nodejs installer. Folks on SO were recommending to use
"lsbom /var/db/receipts/$package | while read file; rm $file; end" (more
or less).

So I had the good idea to look up the manpage of lsbom. In the "See
Also" section, I saw pkgutil and I was like "it might be it!".

But no, it wasn't it. pkgutil can only *install* said packages.

There was, however, a disclaimer stating that "The files and directories
where receipts are stored are subject to change. Always use pkgutil to
query or modify them.". So I took the word and I decided that I'd build
a tool using pkgutil to remove packages.

And this is how pkg-remove was born.