fish: Fix pkg-remove for some packages
You see, I've built pkg-remove to remove just one package. So it was
kind of tailored to this package and as such I assumed that the
identifier of a package always ended with ".pkg".

But, since I've discovered that I can build .pkg bundles myself
(pkgbuild), I've thought that it would be cool to make one...

And I've made a macOS package with a custom identifier, and when I tried
to remove it, it didn't work because the package was called something
like "com.domain.package" and pkg-remove was trying to remove
"com.domain.package.pkg".

So, instead of adding .pkg unconditionally, pkg-remove now adds it only
when the package wasn't found without the ".pkg" extension.