fish: Add token-abbreviations
I guess it is a thing... But I'm fond of it!

Basically, it is like abbreviations but on steroids.
Now, you can choose to expand any token of a command[1] at your liking!

I believe that quite some time ago a friend of mine wanted such a
feature in fish, and it has always stuck in the back of my head. His use
case was the git abbreviations, and I must say that it rocks. In fact,
with any kind of command that takes subcommand, I think it rocks.

[1]: you have to specify its position in the commandline, which is a bit
     weak, and I'm wondering if like "select-completion" wouldn't be
     better. Like a way to tell which completion you'd like to expand
     when you press <space>.