fish: setup: Improve user experience
Now, the setup function doesn't prompt if it should redo the setup if
there are no arguments or the `--help` is specified, because in both
cases, this results in the usage being printed.

Also, the setup.fish script only prints the filename, not the full path
to the script, which is - in my opinion - way prettier. And because
these are my dotfiles, only my opinions matter.