fish: Avoid double expanding of abbrs on enter
So I'm using this custom expand-token function to expand abbreviations
and other tokens in lieu of the built-in mechanism. The problem is that
the built-in expanding takes place when executing the commandline, not
when hitting <return>, so when expand-token does `commandline -f execute`,
the abbreviations are expanded one more time!

That's why I need a way to tell expand-abbr to skip this to let fish
handle it.

(background: I have abbr'ed "n" -> "nvim" and "vim" -> "n" because I try
to switch to n and loose the old habit of using vim but there is a
double expansion to it goes "vim" -> "n" and "n" -> "nvim", which
*cancels* the whole thing...)