nvim: Edit path of every plugin with :Runed
Silly but fix, but quite tricky to find: in neovim (and vim8), the
'runtimepath' gets longer really fast with all of the plugins under
pack/*/start so these plugins aren't displayed in the 'runtimepath'
option. To get a list of the true runtime paths, I have to use
nvim_list_runtime_paths().