nvim: Change `fd` mapping to `fq`
That's actually quite a big change in only three lines. The fundamental
way I interact with neovim is by always switching from normal mode to
insert mode and vice versa. So I've got to choose a good mapping for
this!

Turns out that, too bad, `fd` isn't a great mapping for this. For
instance, I can't type the word `fd` (duh), which is the abbreviation
for "file descriptor" and that's quite annoying.

Now, take `fq`. How often do you type `fq` in a sentence? Because until
now, I had never typed it on purpose (think about it: I had to type `fq`
in this commit message). Thus, `fq` is a *way* better mapping than `fd`.

By the way: this time, I've thought up this change, so I've had these
mappings in my dotfiles for something like a month already and I can
confirm that `fq` is much better than `fd`!