nvim: Add mappings to change font size in GUI
Storing the font spec in a global variable was a good idea because I
don't have to parse the 'guifont' option to get the current font.
This way, it's as simple as updating a variable.

PS: I've amended this commit because I'm not smart enough to know how to
use git...