nvim: Move `opt` plugins to `start` plugins
The reason I'm doing that is that it's just *so much faster*. And it
makes sense: `:packadd` has to go through all the `packpath` to find
where the plugin lies. And this for every plugin!

On the other hand, the `start` plugins go through the `packpath` only
once and add all the plugins in each directory at the same time.