nvim: Remove custom root_dir for sumneko lua server
I don't know why I introduced the custom root_dir in the first place
because with hindsight, it didn't solve any problem, and even worse: it
was causing multiple instances of sumneko[^1] to be spawned when one was
enough (when editing `init.lua`, `plugin/*.lua` or `lua/*.lua`).

[^1]: I know that sumneko isn't the name of the server, just the
(github) name of its creator, but it is IMO better than
lua-language-server anyway.