bin: git switch-remote
This is a script that I've wanted for a long time but I had never
actually done it. This time, however, I sat down and wrote the code.

At first, I was like "oh, it will just be a couple of string replace!".
Turns out that no. Because of myself.

I chose to use a regexp. But being a perfectionist, I obviously wanted
to have a *reliable* regexp, and more importantly, a regexp that
conforms to the standard. So I found the RFC that defined URLs and I
based my regexp on it. Obviously.

I painfully transformed the BNF grammar into several regexps (one regexp
per rule) and the resulting regexp was *massive*. Like 270 characters
long.

I had to simplify it.

I collapsed several layers of rules (to avoid having to use "(?:)" for
each rule) and simplified some things (see the comment before the
regexp) and I've managed to reduce it to 180 chars! (yes that's a lot)

Anyways, using a rule-like system with regexp has turned out to work
quite well, and without it, I couldn't even have made this regexp. (180
chars!)

The rest was then a piece of cake compared to the regexp.