scripts: pkg-remove: Stream file lists
When I wrote the initial implementation, I was already worried about
capturing the output of the command. This stayed in the back of my head
until yesterday, when I decided to make a change.

I benchmarked the non-streaming version and the streaming version, to
eventually find that the streaming version was *faster* (even though it
seemed it was the contrary looking at the verbose output).

The ultimate improvement on this could be using xargs or GNU parallel to
try to squeeze even more performance out of this...